+++ NEWS.md
@@ -47,6 +47,7 @@
 * Fix updateIdentifierTypeCanceledLCCN migration script ([MODINVSTOR-1361](https://folio-org.atlassian.net/browse/MODINVSTOR-1361))
 * Fix inconsistent "messagesPublished" updates caused by asynchronous execution in IterationJobRunner. ([MODINVSTOR-1450](https://folio-org.atlassian.net/browse/MODINVSTOR-1450))
 * Fix items not being updated on /holdings-storage/batch/synchronous ([MODINVSTOR-1368](https://folio-org.atlassian.net/browse/MODINVSTOR-1368))
+* Fix typo in Instance Notes type name. ([MODINVSTOR-1396](https://folio-org.atlassian.net/browse/MODINVSTOR-1396))
 
 ### Tech Dept
 * Stabilize `canRequestOaiPmhViewWithOrderedElectronicAccess` test by refining item lookup logic ([MODINVSTOR-1406](https://folio-org.atlassian.net/browse/MODINVSTOR-1406))